# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Department {
  id: number;
  name: string;
  code: string;
  description?: string | null;
  studentCount?: number;
  courseCount?: number;
}

export interface DepartmentInput {
  name: string;
  code: string;
  description?: string | null;
}

export interface Course {
  id: number;
  name: string;
  code: string;
  departmentId: number;
  departmentName?: string;
  credits: number;
  studentCount?: number;
}

export interface CourseInput {
  name: string;
  code: string;
  departmentId: number;
  credits: number;
}

export type StudentGender = typeof StudentGender[keyof typeof StudentGender];


export const StudentGender = {
  male: 'male',
  female: 'female',
  other: 'other',
} as const;

export interface Student {
  id: number;
  studentCode: string;
  fullName: string;
  email: string;
  phone: string;
  gender: StudentGender;
  dateOfBirth: string;
  address: string;
  departmentId: number;
  departmentName?: string;
  courseId: number;
  courseName?: string;
  year: number;
  semester: number;
  cgpa: number;
  attendancePercentage: number;
  createdAt: string;
}

export type StudentInputGender = typeof StudentInputGender[keyof typeof StudentInputGender];


export const StudentInputGender = {
  male: 'male',
  female: 'female',
  other: 'other',
} as const;

export interface StudentInput {
  studentCode: string;
  fullName: string;
  email: string;
  phone: string;
  gender: StudentInputGender;
  dateOfBirth: string;
  address: string;
  departmentId: number;
  courseId: number;
  year: number;
  semester: number;
  cgpa: number;
  attendancePercentage: number;
}

export interface StudentListResponse {
  items: Student[];
  total: number;
  page: number;
  pageSize: number;
}

export type DashboardSummaryDepartmentDistributionItem = {
  departmentName: string;
  count: number;
};

export type DashboardSummaryYearDistributionItem = {
  year: number;
  count: number;
};

export interface DashboardSummary {
  totalStudents: number;
  totalDepartments: number;
  totalCourses: number;
  averageCgpa: number;
  averageAttendance: number;
  departmentDistribution: DashboardSummaryDepartmentDistributionItem[];
  yearDistribution: DashboardSummaryYearDistributionItem[];
  recentStudents: Student[];
}

export type ListCoursesParams = {
departmentId?: number;
};

export type ListStudentsParams = {
search?: string;
departmentId?: number;
courseId?: number;
year?: number;
page?: number;
pageSize?: number;
};