# lib/api-zod/src/generated/types/dashboardSummary.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { DashboardSummaryDepartmentDistributionItem } from './dashboardSummaryDepartmentDistributionItem';
import type { DashboardSummaryYearDistributionItem } from './dashboardSummaryYearDistributionItem';
import type { Student } from './student';

export interface DashboardSummary {
  totalStudents: number;
  totalDepartments: number;
  totalCourses: number;
  averageCgpa: number;
  averageAttendance: number;
  departmentDistribution: DashboardSummaryDepartmentDistributionItem[];
  yearDistribution: DashboardSummaryYearDistributionItem[];
  recentStudents: Student[];
}