# lib/api-zod/src/generated/types/department.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Department {
  id: number;
  name: string;
  code: string;
  description?: string | null;
  studentCount?: number;
  courseCount?: number;
}