# lib/api-zod/src/generated/types/departmentInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface DepartmentInput {
  name: string;
  code: string;
  description?: string | null;
}