# lib/api-zod/src/generated/types/student.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudentGender } from './studentGender';

export interface Student {
  id: number;
  studentCode: string;
  fullName: string;
  email: string;
  phone: string;
  gender: StudentGender;
  dateOfBirth: Date;
  address: string;
  departmentId: number;
  departmentName?: string;
  courseId: number;
  courseName?: string;
  year: number;
  semester: number;
  cgpa: number;
  attendancePercentage: number;
  createdAt: Date;
}