# lib/api-zod/src/generated/types/studentInput.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { StudentInputGender } from './studentInputGender';

export interface StudentInput {
  studentCode: string;
  fullName: string;
  email: string;
  phone: string;
  gender: StudentInputGender;
  dateOfBirth: Date;
  address: string;
  departmentId: number;
  courseId: number;
  year: number;
  semester: number;
  cgpa: number;
  attendancePercentage: number;
}