# lib/api-zod/src/generated/types/studentListResponse.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { Student } from './student';

export interface StudentListResponse {
  items: Student[];
  total: number;
  page: number;
  pageSize: number;
}